fix(execexpr): ARRAY[] of >=65536 elements silently returned an empty array
ArrayExprStep.nelems was u16 and compile.rs stored `nelems as u16`, so an
ARRAY[] literal with 65536 elements wrapped to 0 and evaluated to an EMPTY
array with NO error raised:

    select array_length(ARRAY[ ...65536 elements... ], 1)
      C:      65536
      pgrust: NULL

65535 was correct, which is what made it hide. C's execExpr.h keeps this count
in an `int` — there is no 16-bit ceiling to honour — so the fix is to widen,
not to add a limit. Found by pgjdbc's PreparedStatement64KBindsTest and
reproduced in bare psql, so it is an engine bug, not a driver artifact.

Silently wrong results are the worst failure class: a crash is loud, this
just hands back the wrong answer. Two siblings carried the same u16 for the
same kind of unbounded count and are widened with it:
  * RowExprStep.nelems  (C rowexpr.nelems is `int`)
  * MinMax.nelems       (GREATEST/LEAST over an int-length arg List)
Audited the rest: the remaining `.len() as u16` sites are WAL-record and
page-item counts (nbtree, hash, gist, heap), which are genuinely uint16 in C's
WAL structs and bounded by page size — those are correct.

Regression test walks the boundary at 65535/65536/65537, asserting both the
count the compiled step carries and the length of the array actually built.
Proven non-vacuous: reintroducing `as u16` makes it fail with "wrapped at
65536 elements".

Also in this commit, the elog source-line fixes from the same driver sweep:
NoticeResponse/ErrorResponse now always emit PG_DIAG_SOURCE_LINE when a
location is present, because C's ereport macros always pass __LINE__ and
clients rely on it — asyncpg does an unguarded del dct['server_source_line'],
which raised KeyError inside its notice callback and hung the connection. A 0
there means "this ported site has no C line yet" (greppable) rather than a
fabricated number; pl/pgsql RAISE now carries its real one, pl_exec.c:3909,
verified against the REL_18_3 source.

+++ crates/backend/executor/execexpr/src/compile.rs
@@ -2529,7 +2529,7 @@ fn init_array_expr<'mcx>(
 
     Ok(Step::ArrayExprStep {
         elems,
-        nelems: nelems as u16,
+        nelems: nelems as u32,
         frame: frame_ix,
         elmtype: arr.element_typeid,
         elmlen,
@@ -3303,7 +3303,7 @@ fn init_row_expr<'mcx>(
             .write(core::mem::transmute::<TupleDescData<'mcx>, TupleDescData<'static>>(desc));
     }
 
-    Ok(Step::RowExprStep { elems, nelems: nelems as u16, frame: frame_ix, desc: desc_ptr, out })
+    Ok(Step::RowExprStep { elems, nelems: nelems as u32, frame: frame_ix, desc: desc_ptr, out })
 }
 
 // C ExecInitExprRec T_JsonConstructorExpr (execExpr.c:2379): args evaluate
@@ -4006,7 +4006,7 @@ fn init_minmax<'mcx>(
     Ok(Step::MinMax {
         call,
         slots,
-        nelems: nelems as u16,
+        nelems: nelems as u32,
         least: mm.op == ::types_nodes::primnodes::MinMaxOp::IS_LEAST,
         out,
     })
